wrote some unit tests as an example of what's been refactored
wrote some unit tests as an example of what's been refactored
